Add an exception avoidance code in true.cs
In .NET, when using dictionary item[] property, If the target key is not
exist, it throws exception. In trie.cs, a nullable Node cur2 variable
was initialized by just a item property. Replace it to a conditional
statement which uses TryGetValue's returned value. When it returns
false, returns cur and exit the current function call.